Fixing an NPE when cleaning up some test studies

+++ app/org/sagebionetworks/bridge/services/UserAdminServiceImpl.java
@@ -141,9 +141,10 @@ public class UserAdminServiceImpl implements UserAdminService {
         // This user may have never consented to research. Ignore if that's the case.
         for (String trackerId : study.getTrackers()) {
             Tracker tracker = studyService.getTrackerByIdentifier(trackerId);
-            
-            HealthDataKey key = new HealthDataKey(study, tracker, user);
-            healthDataService.deleteHealthDataRecords(key);
+            if (tracker != null) { // this happens with some tests
+                HealthDataKey key = new HealthDataKey(study, tracker, user);
+                healthDataService.deleteHealthDataRecords(key);
+            }
         }
     }
